Ajustar codigo en .jsx separados
