+++ package.json
@@ -4,8 +4,8 @@
     "description": "",
     "main": "index.js",
     "scripts": {
-        "compile": "babel src --presets react,es2015 --out-dir static",
-        "watch": "babel src --presets react,es2015 --out-dir static --watch",
+        "compile": "webpack",
+        "watch": "webpack --watch",
         "test": "echo \"Error: no test specified\" && exit 1",
         "start": "nodemon -w server server/server.js"
     },
@@ -20,6 +20,7 @@
     },
     "devDependencies": {
         "babel-cli": "^6.26.0",
+        "babel-loader": "^7.1.5",
         "babel-preset-es2015": "^6.24.1",
         "babel-preset-react": "^6.24.1",
         "nodemon": "^2.0.2",
