+++ src/IssueFilter.jsx
@@ -1,0 +1,5 @@
+export default class IssueFilter extends React.Component {
+  render() {
+    return <div>Marcador para el filtro de problemas (issue's)</div>;
+  }
+}
